chore: format code changes

+++ apps/api/plane/app/views/project/base.py
@@ -55,7 +55,9 @@ class ProjectViewSet(BaseViewSet):
             super()
             .get_queryset()
             .filter(workspace__slug=self.kwargs.get("slug"))
-            .select_related("workspace", "workspace__owner", "default_assignee", "project_lead")
+            .select_related(
+                "workspace", "workspace__owner", "default_assignee", "project_lead"
+            )
             .annotate(
                 is_favorite=Exists(
                     UserFavorite.objects.filter(
@@ -93,15 +95,17 @@ class ProjectViewSet(BaseViewSet):
             .distinct()
         )
 
-    @allow_permission(allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE")
+    @allow_permission(
+        allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE"
+    )
     def list_detail(self, request, slug):
         fields = [field for field in request.GET.get("fields", "").split(",") if field]
         projects = self.get_queryset().order_by("sort_order", "name")
         if WorkspaceMember.objects.filter(
-                member=request.user,
-                workspace__slug=slug,
-                is_active=True,
-                role=ROLE.GUEST.value,
+            member=request.user,
+            workspace__slug=slug,
+            is_active=True,
+            role=ROLE.GUEST.value,
         ).exists():
             projects = projects.filter(
                 project_projectmember__member=self.request.user,
@@ -109,10 +113,10 @@ class ProjectViewSet(BaseViewSet):
             )
 
         if WorkspaceMember.objects.filter(
-                member=request.user,
-                workspace__slug=slug,
-                is_active=True,
-                role=ROLE.MEMBER.value,
+            member=request.user,
+            workspace__slug=slug,
+            is_active=True,
+            role=ROLE.MEMBER.value,
         ).exists():
             projects = projects.filter(
                 Q(
@@ -127,13 +131,19 @@ class ProjectViewSet(BaseViewSet):
                 order_by=request.GET.get("order_by", "-created_at"),
                 request=request,
                 queryset=(projects),
-                on_results=lambda projects: ProjectListSerializer(projects, many=True).data,
+                on_results=lambda projects: ProjectListSerializer(
+                    projects, many=True
+                ).data,
             )
 
-        projects = ProjectListSerializer(projects, many=True, fields=fields if fields else None).data
+        projects = ProjectListSerializer(
+            projects, many=True, fields=fields if fields else None
+        ).data
         return Response(projects, status=status.HTTP_200_OK)
 
-    @allow_permission(allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE")
+    @allow_permission(
+        allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE"
+    )
     def list(self, request, slug):
         sort_order = ProjectMember.objects.filter(
             member=self.request.user,
@@ -144,7 +154,9 @@ class ProjectViewSet(BaseViewSet):
 
         projects = (
             Project.objects.filter(workspace__slug=self.kwargs.get("slug"))
-            .select_related("workspace", "workspace__owner", "default_assignee", "project_lead")
+            .select_related(
+                "workspace", "workspace__owner", "default_assignee", "project_lead"
+            )
             .annotate(
                 member_role=ProjectMember.objects.filter(
                     project_id=OuterRef("pk"),
@@ -179,10 +191,10 @@ class ProjectViewSet(BaseViewSet):
         )
 
         if WorkspaceMember.objects.filter(
-                member=request.user,
-                workspace__slug=slug,
-                is_active=True,
-                role=ROLE.GUEST.value,
+            member=request.user,
+            workspace__slug=slug,
+            is_active=True,
+            role=ROLE.GUEST.value,
         ).exists():
             projects = projects.filter(
                 project_projectmember__member=self.request.user,
@@ -190,10 +202,10 @@ class ProjectViewSet(BaseViewSet):
             )
 
         if WorkspaceMember.objects.filter(
-                member=request.user,
-                workspace__slug=slug,
-                is_active=True,
-                role=ROLE.MEMBER.value,
+            member=request.user,
+            workspace__slug=slug,
+            is_active=True,
+            role=ROLE.MEMBER.value,
         ).exists():
             projects = projects.filter(
                 Q(
@@ -204,14 +216,22 @@ class ProjectViewSet(BaseViewSet):
             )
         return Response(projects, status=status.HTTP_200_OK)
 
-    @allow_permission(allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE")
+    @allow_permission(
+        allowed_roles=[ROLE.ADMIN, ROLE.MEMBER, ROLE.GUEST], level="WORKSPACE"
+    )
     def retrieve(self, request, slug, pk):
-        project = self.get_queryset().filter(archived_at__isnull=True).filter(pk=pk).first()
+        project = (
+            self.get_queryset().filter(archived_at__isnull=True).filter(pk=pk).first()
+        )
 
         if project is None:
-            return Response({"error": "Project does not exist"}, status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                {"error": "Project does not exist"}, status=status.HTTP_404_NOT_FOUND
+            )
 
-        member_ids = [str(project_member.member_id) for project_member in project.members_list]
+        member_ids = [
+            str(project_member.member_id) for project_member in project.members_list
+        ]
 
         if str(request.user.id) not in member_ids:
             if project.network == ProjectNetwork.SECRET.value:
@@ -242,8 +262,12 @@ class ProjectViewSet(BaseViewSet):
         is_free_workspace = workspace.plan == WorkspacePlan.FREE.value
 
         if is_free_workspace:
-            raise PermissionDenied("Cannot create more than one project in a free workspace.")
-        serializer = ProjectSerializer(data={**request.data}, context={"workspace_id": workspace.id})
+            raise PermissionDenied(
+                "Cannot create more than one project in a free workspace."
+            )
+        serializer = ProjectSerializer(
+            data={**request.data}, context={"workspace_id": workspace.id}
+        )
         if serializer.is_valid():
             serializer.save()
 
@@ -254,11 +278,13 @@ class ProjectViewSet(BaseViewSet):
                 role=ROLE.ADMIN.value,
             )
             # Also create the issue property for the user
-            _ = IssueUserProperty.objects.create(project_id=serializer.data["id"], user=request.user)
+            _ = IssueUserProperty.objects.create(
+                project_id=serializer.data["id"], user=request.user
+            )
 
-            if serializer.data["project_lead"] is not None and str(serializer.data["project_lead"]) != str(
-                    request.user.id
-            ):
+            if serializer.data["project_lead"] is not None and str(
+                serializer.data["project_lead"]
+            ) != str(request.user.id):
                 ProjectMember.objects.create(
                     project_id=serializer.data["id"],
                     member_id=serializer.data["project_lead"],
@@ -331,7 +357,9 @@ class ProjectViewSet(BaseViewSet):
 
         project = Project.objects.get(pk=pk)
         intake_view = request.data.get("inbox_view", project.intake_view)
-        current_instance = json.dumps(ProjectSerializer(project).data, cls=DjangoJSONEncoder)
+        current_instance = json.dumps(
+            ProjectSerializer(project).data, cls=DjangoJSONEncoder
+        )
         if project.archived_at:
             return Response(
                 {"error": "Archived projects cannot be updated"},
@@ -373,19 +401,19 @@ class ProjectViewSet(BaseViewSet):
 
     def destroy(self, request, slug, pk):
         if (
-                WorkspaceMember.objects.filter(
-                    member=request.user,
-                    workspace__slug=slug,
-                    is_active=True,
-                    role=ROLE.ADMIN.value,
-                ).exists()
-                or ProjectMember.objects.filter(
-            member=request.user,
-            workspace__slug=slug,
-            project_id=pk,
-            role=ROLE.ADMIN.value,
-            is_active=True,
-        ).exists()
+            WorkspaceMember.objects.filter(
+                member=request.user,
+                workspace__slug=slug,
+                is_active=True,
+                role=ROLE.ADMIN.value,
+            ).exists()
+            or ProjectMember.objects.filter(
+                member=request.user,
+                workspace__slug=slug,
+                project_id=pk,
+                role=ROLE.ADMIN.value,
+                is_active=True,
+            ).exists()
         ):
             project = Project.objects.get(pk=pk, workspace__slug=slug)
             project.delete()
@@ -423,7 +451,9 @@ class ProjectArchiveUnarchiveEndpoint(BaseAPIView):
         project.archived_at = timezone.now()
         project.save()
         UserFavorite.objects.filter(workspace__slug=slug, project=project_id).delete()
-        return Response({"archived_at": str(project.archived_at)}, status=status.HTTP_200_OK)
+        return Response(
+            {"archived_at": str(project.archived_at)}, status=status.HTTP_200_OK
+        )
 
     @allow_permission([ROLE.ADMIN, ROLE.MEMBER])
     def delete(self, request, slug, project_id):
@@ -439,18 +469,26 @@ class ProjectIdentifierEndpoint(BaseAPIView):
         name = request.GET.get("name", "").strip().upper()
 
         if name == "":
-            return Response({"error": "Name is required"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response(
+                {"error": "Name is required"}, status=status.HTTP_400_BAD_REQUEST
+            )
 
-        exists = ProjectIdentifier.objects.filter(name=name, workspace__slug=slug).values("id", "name", "project")
+        exists = ProjectIdentifier.objects.filter(
+            name=name, workspace__slug=slug
+        ).values("id", "name", "project")
 
-        return Response({"exists": len(exists), "identifiers": exists}, status=status.HTTP_200_OK)
+        return Response(
+            {"exists": len(exists), "identifiers": exists}, status=status.HTTP_200_OK
+        )
 
     @allow_permission([ROLE.ADMIN, ROLE.MEMBER], level="WORKSPACE")
     def delete(self, request, slug):
         name = request.data.get("name", "").strip().upper()
 
         if name == "":
-            return Response({"error": "Name is required"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response(
+                {"error": "Name is required"}, status=status.HTTP_400_BAD_REQUEST
+            )
 
         if Project.objects.filter(identifier=name, workspace__slug=slug).exists():
             return Response(
@@ -467,7 +505,9 @@ class ProjectUserViewsEndpoint(BaseAPIView):
     def post(self, request, slug, project_id):
         project = Project.objects.get(pk=project_id, workspace__slug=slug)
 
-        project_member = ProjectMember.objects.filter(member=request.user, project=project, is_active=True).first()
+        project_member = ProjectMember.objects.filter(
+            member=request.user, project=project, is_active=True
+        ).first()
 
         if project_member is None:
             return Response({"error": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
@@ -496,7 +536,9 @@ class ProjectFavoritesViewSet(BaseViewSet):
             .get_queryset()
             .filter(workspace__slug=self.kwargs.get("slug"))
             .filter(user=self.request.user)
-            .select_related("project", "project__project_lead", "project__default_assignee")
+            .select_related(
+                "project", "project__project_lead", "project__default_assignee"
+            )
             .select_related("workspace", "workspace__owner")
         )
